Add files via upload
Here is the Arduino code for the Revox control. Sorry it looks so clunky. Maybe someone can streamline it for me... But this works for me.

I have also attached an Excel file that shows the MIDI on notes and functions. Depending on your hardware, you may need to add a MIDI note off message to your sequence. I did it via my MIDI translator software- Bome Midi Translator Pro.